Reverting change to use NetworkController in the NetworkConnector rather than NetworkChannel.
While using a NetworkController has it's uses, allowing the NetworkConnector to easily register with the controller, however it lets the connector do things we won't want it to do. However we will have to find an alternative way of registering, http://sammacbeth.lighthouseapp.com/projects/65108/tickets/26-network-registration proposes a possible fix.